Backend/requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
get==2019.4.13
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
post==2019.4.13
public==2019.4.13
query-string==2019.4.13
request==2019.4.13
six==1.12.0
Werkzeug==0.15.6
bcrypt==3.1.7
